+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * JSON-RPC
   * Changed hardcoded HTTPS ports to parameters (PR#23)
   * Fixed a bug of eth_getBlockByNumber
+  * Implement eth_gasPrice
+* Command
+  * `smartbchd start` will not ignore `--home` option
+  * `smartbchd start` can disable HTTPS-RPC and WSS-RPC server now
 
 
 ## v0.3.3